Rename descriptor_patching → program_descriptor_patching
